% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{cluster_scatter}
\alias{cluster_scatter}
\title{Scatter plot for genes}
\usage{
cluster_scatter(
  dat,
  mapping = "umap",
  group = NULL,
  size = 1,
  text = T,
  text_group = NULL,
  text_color = "black",
  text_size = 5,
  segment_size = 0.75,
  text_face = "bold"
)
}
\arguments{
\item{dat}{Seurat/monocle object.}

\item{mapping}{Name of reduction to pull cell embeddings for. Only for Seurat object.}

\item{group}{Factor. Grouping information of each gene. Must contain the same length of genes.}

\item{size}{Numeric. The point size of scatter plot. Default:1.}

\item{text}{Logical. If TRUE, gene family will be labeled at scatter plot. Default:T.}

\item{text_group}{Factor. Grouping information of labeled gene family. Must contain the same length of genes.}

\item{text_color}{Character. Colors of labeled gene family. Default:"black".}

\item{text_size}{Numeric. The gene family text size of scatter plot. Default:5.}

\item{segment_size}{Numeric. The segment size of scatter plot. Default:0.75.}

\item{text_face}{Character. fontface of labeled gene family.}
}
\value{
Scatter plot for genes.
}
\description{
Scatter plot for alignment result after dimensionality reduction.
}
