% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{combine_alignment}
\alias{combine_alignment}
\title{Combine multiple alignment results according to the distance of each alignment result.}
\usage{
combine_alignment(
  alignment_list = NULL,
  alignment_file = NULL,
  group_dist = NULL,
  type = "AA",
  tree = F,
  ...
)
}
\arguments{
\item{alignment_list}{A list containing multiple alignment results.}

\item{alignment_file}{A vector containing paths of multiple alignment results.}

\item{group_dist}{Dist. This parameter could be applied to import the calculated distance between groups. If NULL, distance between groups will be calculated automatically.}

\item{type}{Sequences type. One of DNA, RNA and AA.}

\item{tree}{Logical. If TRUE, also return hclust tree of each family. Default: FALSE.}

\item{...}{Other parameter of group_alignment}
}
\value{
multiple alignment results with hclust tree
}
\description{
Combine multiple alignment results according to the distance of each alignment result.
}
