% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{constructingNetwork}
\alias{constructingNetwork}
\title{Getting gene lineage from score matrix.}
\usage{
constructingNetwork(
  data,
  range = NULL,
  thresh = 0.1,
  normalization = T,
  group = NULL,
  alpha = 0.01,
  theta1 = 0.8,
  rootNode = 1
)
}
\arguments{
\item{data}{Matrix. Score matrix of each site.}

\item{range}{Numeric. Extract the specified column of the score matrix for network construction.}

\item{thresh}{Numeric. Edges whose correlation is less than thresh will be filtered out.}

\item{normalization}{Logical. To normalize the data or not. Default:TRUE.}

\item{group}{Factor. Grouping information of each gene. Must contain the same length of nrow(data).}

\item{alpha}{significance level for a two-sided Wilcoxon rank-sum test of the distribution of scEnergy between two metagenes.Default:0.01.}

\item{theta1}{Numeric. Select the genes which include theta1 energy in each cluster. Default:0.8.}

\item{rootNode}{Numeric. The root node (i.e., start state) for inferring gene family lineages.}
}
\value{
a struct giving the gene lineage information.
}
\description{
Getting gene lineage from score matrix.
}
