% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{edmonds}
\alias{edmonds}
\title{Getting circuit with min dist.}
\usage{
edmonds(V, E, rootNode)
}
\arguments{
\item{V}{Nodes of graphs.}

\item{E}{Edge of graphs. Containing weights.}

\item{rootNode}{Numeric. Rootnode of graphs.}
}
\value{
circuit with min dist.
}
\description{
Getting circuit with min dist.
}
