% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{find_ckt_edges_to_add}
\alias{find_ckt_edges_to_add}
\title{Getting circuit with min dist.}
\usage{
find_ckt_edges_to_add(ED, level, RECONBUCKET)
}
\arguments{
\item{ED}{List contain graphical structure.}

\item{level}{Number of iterations}

\item{RECONBUCKET}{edge in circuit}
}
\description{
Getting circuit with min dist.
}
