% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{find_Specific_site}
\alias{find_Specific_site}
\title{Find specific sites for each gene family.}
\usage{
find_Specific_site(
  seurat,
  ident = NULL,
  only.pos = TRUE,
  logfc = 0.1,
  min.pct = 0,
  ...
)
}
\arguments{
\item{seurat}{Seurat object.}

\item{ident}{Character. One of colnames of Seurat object metadata. If NULL, use the active.ident of the Seurat object as the grouping information.}

\item{only.pos}{Only return positive marker site. Default:TRUE.}

\item{logfc}{Limit testing to genes which show, on average, at least X-fold difference (log-scale) between the two groups of cells. Default:1.}

\item{min.pct}{only test genes that are detected in a minimum fraction of min.pct cells in either of the two populations. Meant to speed up the function by not testing genes that are very infrequently expressed. Default: 0.25.}

\item{...}{Other parameters of Seurat::FindAllMarkers.}
}
\value{
Matrix containing a ranked list of putative markers, and associated statistics (p-values, ROC score, etc.)
}
\description{
Find specific sites for each gene family.
}
\seealso{
\code{\link{FindAllMarkers}}
}
