% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{finde0e1}
\alias{finde0e1}
\title{Extract the weight of the end node of edges which are in circuit.}
\usage{
finde0e1(ED, EP)
}
\arguments{
\item{ED}{List contain graphical structure.}

\item{EP}{Numeric. The end node number in graphs.}
}
\value{
the weight of the end node of edges which are in circuit.
}
\description{
Extract the weight of the end node of edges which are in circuit.
}
