% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alignment_score.R
\name{get_alignment_score}
\alias{get_alignment_score}
\title{Calculate score of each site in alignment}
\usage{
get_alignment_score(
  alignment = NULL,
  alignment_file = NULL,
  standard = F,
  symbol = "site",
  type = c("DNA", "RNA", "AA"),
  seq_length_limit = 0L,
  remove_duplicated = F
)
}
\arguments{
\item{alignment}{input sequence alignment result, for DNA/RNA/protein sequences, DNAStringSet/RNAStringSet/AAStringSet or character class were required. All sequences must have the same length.}

\item{alignment_file}{the path of input sequence alignment result.}

\item{standard}{Logical. If TRUE, Only keep the score of AA_STANDARD/DNA_BASES/RNA_BASES for protein/DNA/RNA sequences. Default:FALSE.}

\item{symbol}{Character. Prefix of colnames of result. Default:"site".}

\item{type}{Input sequence type.One of "DNA", "RNA" or "AA".}

\item{seq_length_limit}{Integer. Lower limit of input sequence length. Default:0.}

\item{remove_duplicated}{Logical. If TRUE, only one sequence with repeated names in the score matrix is kept. Default:FALSE.}
}
\value{
score matrix of each site. rownames of the repeated sequence name will be suffixed.
}
\description{
Calculate score of each site in alignment
}
\examples{
require(ggseqlogo)
data(ggseqlogo_sample)
test_aa <- get_alignment_score(alignment = seqs_aa[[1]],type = "AA")
test_dna <- get_alignment_score(alignment = seqs_dna[[1]],type = "DNA")
}
