% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{get_bits}
\alias{get_bits}
\title{Get bits of each site.}
\usage{
get_bits(seq_prob, type = c("DNA", "RNA", "AA"), freq = T, base = T, prob = F)
}
\arguments{
\item{seq_prob}{Seurat object.}

\item{type}{Character. One of DNA, RNA and AA.}

\item{freq}{Logical. If TRUE, the final result will be multiplied by the ratio of valid characters. Default:TRUE.}

\item{base}{Logical. If TRUE, only keep the score of AA_STANDARD/DNA_BASES/RNA_BASES for protein/DNA/RNA sequences. Default:TRUE.}

\item{prob}{Logical. If TRUE, return seq prob of each site. Default:FALSE.}
}
\value{
bits/seq prob of each site
}
\description{
Get bits of each site.
}
