% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{get_seq_percent}
\alias{get_seq_percent}
\title{Calculate the Percent matrix of each site.}
\usage{
get_seq_percent(dat, alphabet)
}
\arguments{
\item{dat}{score matrix.}

\item{alphabet}{alphabet of dat.}
}
\value{
Percent matrix of each site for each element in alphabet.
}
\description{
Calculate the Percent matrix of each site.
}
