% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Homeobox_align}
\alias{Homeobox_align}
\title{Multiple sequences alignment of homeobox transcription factor superfamily}
\format{
A list with 5 elements:
\describe{
  \item{HB_other}{multiple sequences alignment result of DNA
                  binding domain (DBD) of HB_other,
     \url{http://planttfdb.gao-lab.org/multi_align/family/HB-other/domain_aln.fas}}
  \item{HB_PHD}{multiple sequences alignment result of DNA
                  binding domain (DBD) of HB_PHD,
     \url{http://planttfdb.gao-lab.org/multi_align/family/HB-PHD/domain_aln.fas}}
  \item{HD_ZIP}{multiple sequences alignment result of DNA
                  binding domain (DBD) of HD_ZIP,
     \url{http://planttfdb.gao-lab.org/multi_align/family/HD_ZIP/domain_aln.fas}}
  \item{TALE}{multiple sequences alignment result of DNA
                  binding domain (DBD) of TALE,
     \url{http://planttfdb.gao-lab.org/multi_align/family/TALE/domain_aln.fas}}
  \item{WOX}{multiple sequences alignment result of DNA
                  binding domain (DBD) of WOX,
     \url{http://planttfdb.gao-lab.org/multi_align/family/WOX/domain_aln.fas}}
}
}
\usage{
Homeobox_align
}
\description{
A dataset containing first 200 sequences in multiple sequences
alignment result of DNA binding domain (DBD) of HB_other, HB_PHD,
HD_ZIP, TALE and WOX transcription factor family
}
\keyword{datasets}
