% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Homeobox_annot}
\alias{Homeobox_annot}
\title{Annotation of DNA binding domain of homeobox transcription factor superfamily}
\format{
A data frame with 1000 rows and 11 variables:
\describe{
  \item{Species}{Species of transcription factors}
  \item{ID}{Gene id of transcription factors}
  \item{Names}{Names of DNA binding domain}
  \item{Family}{Family of transcription factors}
  \item{Description}{Description of transcription factor family}
  \item{Taxonomic1}{Phylum corresponding to each species}
  \item{Taxonomic2}{Phylum or Subclass corresponding to each species}
  \item{Alias}{Alias of species}
  \item{Genome}{Species with or without genome sequence,
    \url{http://planttfdb.gao-lab.org/index.php}}
  \item{WOX_subgroup}{Subfamily of WOX transcription factors family}
  \item{At_WOX}{The most similar WOX gene in Arabidopsis, only for TF of WOX family}
}
}
\usage{
Homeobox_annot
}
\description{
Annotation of DNA binding domain of homeobox transcription factor superfamily
}
\keyword{datasets}
