% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{import_metadata}
\alias{import_metadata}
\title{import metadata from data frame}
\usage{
import_metadata(seurat, metadata)
}
\arguments{
\item{seurat}{Seurat object.}

\item{metadata}{Number of features to select as top variable features.}
}
\value{
Seurat object with imported metadata
}
\description{
import metadata from data frame
}
