% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{index_of_max_value_incoming_edge}
\alias{index_of_max_value_incoming_edge}
\title{Get max value incoming edge and its index from edge matrix.}
\usage{
index_of_max_value_incoming_edge(G, VERTEX)
}
\arguments{
\item{G}{Edge matrix.}

\item{VERTEX}{The node number in G.}
}
\value{
A list contain EDGE_VALUE and EDGE_ADDED
}
\description{
Get max value incoming edge and its index from edge matrix.
}
