% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{iscycle}
\alias{iscycle}
\title{Get shortest paths from edge matrix.}
\usage{
iscycle(G, S, D)
}
\arguments{
\item{G}{Edge matrix.}

\item{S}{Numeric. The Start node number in G.}

\item{D}{Numeric. The End node number in G.}
}
\value{
A list contain dist and path between Start node number and End node number.
}
\description{
Get shortest paths from edge matrix.
}
