% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{MDST_plot}
\alias{MDST_plot}
\title{MDST Network visualization with igraph}
\usage{
MDST_plot(
  MDST,
  node.color = NULL,
  group = NULL,
  weight = T,
  node.label.color = "black",
  edge.color = "black",
  arrow.size = 0.4,
  node.size = NULL,
  frame.color = NULL,
  edge.label = NULL,
  edge.width = 2,
  edge.label.color = "black",
  edge.label.loc = 0.5,
  edge.label.adjust = 0.1,
  edge.label.cex = 0.8,
  arrow.width = 1.5,
  my_layout = layout_as_tree,
  rescale = F,
  node.self = F,
  ...
)
}
\arguments{
\item{MDST}{MDST Network.}

\item{node.color}{The fill color of the vertex.}

\item{group}{The vertex labels.}

\item{weight}{Logical. Whether to determine the width of the edges. according to the weight.Default:TRUE}

\item{node.label.color}{The color of the node labels.}

\item{edge.color}{The color of the edges.}

\item{arrow.size}{Size of arrows.}

\item{node.size}{Size of nodes.}

\item{frame.color}{The color of the frame of the vertices.}

\item{edge.label}{The edge labels.}

\item{edge.width}{Width of each edge.}

\item{edge.label.color}{The color of the edge labels.}

\item{edge.label.loc}{The location of edge labels, if edge.label.loc = 0, edge label locate at the same position of start node,
if edge.label.loc = 1, edge label locate at the same position of end node,
if edge.label.loc = 0.5, edge label locate at the middle of start and end node.Default:0.5.}

\item{edge.label.adjust}{edge label will be adjusted in the x-axis direction.Default:0.1.}

\item{edge.label.cex}{The font size for the edge labels.Default:0.8.}

\item{arrow.width}{The width of the arrows. Default:layout_as_tree}

\item{my_layout}{Either a function or a numeric matrix. It specifies how the vertices will be placed on the plot.}

\item{rescale}{Logical. Whether to rescale the coordinates to the [-1,1]x[-1,1](x[-1,1]) interval.
This parameter is not implemented for tkplot.Default:FALSE, the layout will NOT be rescaled.}

\item{node.self}{Logical. Whether to draw an arrow pointing to the vertice itself.Default:FALSE.}

\item{...}{Other parameter of plot.igraph.}
}
\description{
MDST Network visualization with igraph
}
\seealso{
\code{\link{plot.igraph},\link{igraph.plotting}}
}
