% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{monocle_import}
\alias{monocle_import}
\title{Import a seurat or scatter/scran CellDataSet object and convert it to a monocle cds.}
\usage{
monocle_import(otherCDS, import_all = FALSE)
}
\arguments{
\item{otherCDS}{The object you would like to convert into a monocle cds.}

\item{import_all}{Whether or not to import all the slots in seurat or scatter. Default is FALSE (or only keep minimal dataset).}
}
\value{
a new monocle cell dataset object converted from other objects (Scatter or Seurat).
}
\description{
This function is modified by the \code{\link{importCDS}} function.
}
