% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{monocle_workflow}
\alias{monocle_workflow}
\title{Dimensionality reduction, classification and pseudotime analysis of the score matrix.}
\usage{
monocle_workflow(seurat, method = "DDRTree", ncenter = 5, ...)
}
\arguments{
\item{seurat}{Seurat object.}

\item{method}{A character string specifying the algorithm to use for dimensionality reduction.}

\item{...}{Other parameters of reduceDimension and orderCells functions in monocle package.}
}
\value{
CellDataSet object with dimensionality reduction, classification and pseudotime analysis results.
}
\description{
Dimensionality reduction, classification and pseudotime analysis of the score matrix.
}
\seealso{
\code{\link{reduceDimension},\link{orderCells}}
}
