% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{multiple_group_scatter}
\alias{multiple_group_scatter}
\title{Scatter plot for genes belonging to multiple gene families}
\usage{
multiple_group_scatter(
  dat,
  my_color,
  mapping = "umap",
  group = NULL,
  gene = NULL,
  background_color = "grey80",
  text = T,
  text_group = NULL,
  text_face = "bold",
  text_color = "grey60",
  text_size = 5,
  legend = T,
  point_size = 1,
  segment_size = 1,
  multiple_point_size = 0.3
)
}
\arguments{
\item{dat}{Seurat/monocle object.}

\item{my_color}{Gene families colors. Only gene belonging to multiple gene families will be colored.}

\item{mapping}{Name of reduction to pull gene embedding for. Only for Seurat object.}

\item{group}{Factor. Grouping information of each gene. Must contain the same length of genes.}

\item{gene}{Character. Must contain duplicate gene names.}

\item{background_color}{Colors for gene NOT belonging to multiple gene families.}

\item{text}{Logical. If TRUE, gene family will be labeled at scatter plot. Default:T.}

\item{text_group}{Factor. Grouping information of labeled gene family. Must contain the same length of genes.}

\item{text_face}{Character. fontface of labeled gene family.}

\item{text_color}{Character. Colors of labeled gene family. Default:"black".}

\item{text_size}{Numeric. The gene family text size of scatter plot. Default:5.}

\item{legend}{Logical. If TRUE, add a legend to the picture. Default:TRUE.}

\item{point_size}{Numeric. The background point size of scatter plot. Default:1.}

\item{segment_size}{Numeric. The segment size of scatter plot. Only for monocle object. Default:1.}

\item{multiple_point_size}{Numeric. Point size of gene belonging to multiple gene families. Default:0.3.}
}
\value{
Scatter plot for genes belonging to multiple gene families.
}
\description{
Scatter plot for genes belonging to multiple gene families,
Genes belonging to multiple families will be given multiple colors
}
