% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Position_site_heatmap}
\alias{Position_site_heatmap}
\title{Heatmap for specific site of each family}
\usage{
Position_site_heatmap(
  seq_conservation,
  group_color = NULL,
  labs_x = "",
  labs_y = "Gene Family"
)
}
\arguments{
\item{seq_conservation}{Conservation matrix of each site.}

\item{group_color}{Gene families colors.}

\item{labs_x}{Character. Title of X lab.}

\item{labs_y}{Character. Title of Y lab.}
}
\value{
Heatmap for specific site of each family.
}
\description{
Heatmap for specific site of each family
}
