% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Position_site_plot}
\alias{Position_site_plot}
\title{Draw a rectangle to reflect the specific site of each family.}
\usage{
Position_site_plot(
  dat,
  site = "gene",
  cluster = "cluster",
  seurat,
  site_color,
  legend = T,
  plot_margin = margin(5, 5, 50, 5)
)
}
\arguments{
\item{dat}{Data frame. Specific site data.}

\item{site}{Character. Site column in dat. Default:"gene".}

\item{cluster}{Character. Cluster column in dat. Default:"cluster".}

\item{seurat}{Seurat object.}

\item{site_color}{Character. Color of each group.}

\item{legend}{Logical. If TRUE, add a legend to the picture. Default:TRUE.}

\item{plot_margin}{margin around entire plot (unit with the sizes of the
top, right, bottom, and left margins).Default:margin(5,5,50,5)}
}
\value{
A ggplot of a rectangle to reflect the specific site of each family.
}
\description{
Draw a rectangle to reflect the specific site of each family.
}
