% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{position_site_weblogo}
\alias{position_site_weblogo}
\title{weblogo for specific site of each family}
\usage{
position_site_weblogo(
  seqs = NULL,
  alignment_file = NULL,
  group = NULL,
  type = "AA",
  max_seq_length = 30,
  ncol = 1,
  range = "conservation",
  stack_width_threshold = 0.2,
  Specific_site = NULL,
  legend = F,
  stack_width = NULL,
  position_sort = T,
  relative = F,
  mark = F,
  mark_data = NULL,
  title = T,
  title.color = NULL,
  text_angle = 0,
  site_size = 10
)
}
\arguments{
\item{seqs}{Aligned sequences.}

\item{alignment_file}{Path of aligned sequences.}

\item{group}{Factor. Grouping information of each gene. Must contain the same length of genes.}

\item{type}{Character. Sequences type, one of DNA, RNA and AA.}

\item{max_seq_length}{Numeric. The maximum number of site plotted in the weblogo.}

\item{ncol}{Numeric. The number of columns in Weblogo}

\item{range}{If range = "conservation", Automatic selection of the most conservative sites.(default)
If range = "site", Automatic selection of the most conservative sites from Specific_site data.
If class(range) is data.frame, each column in the range represents the locus selected by each family.
If class(range) is list, each element in the range represents the locus selected by each family.
If is.null(range), first max_seq_length site will be selected by each family.
if class(range) is integer/numeric, corresponding site will be selected by each family.}

\item{stack_width_threshold}{Numeric. Stack width lower than stack_width_threshold will be changed to 0.}

\item{Specific_site}{Specific_site data. Only available when range = "site"}

\item{legend}{Logical. If FALSE, legend of weblogo will be deleted. Default: FALSE.}

\item{stack_width}{Letter width in Weblogo. If is.null(stack_width), the letter width will be calculated automatically according to the conservatism of the locus.
If class(stack_width) is data.frame/list, The letter width will be extracted from stack_width.
If stack_width is a length 1 numeric/integer vector, width of all letters are set to stack_width.
If stack_width is a numeric/integer vector with length > 1. The letter width will be recycled.}

\item{position_sort}{Logical. If TRUE, position of sites will be sorted. Default: TRUE.}

\item{relative}{Logical. If FALSE, ylim of weblogo will be fixed.}

\item{mark}{Logical. If TRUE, Mark on Weblogo. Default: FALSE.}

\item{mark_data}{Data.frame or List. Only available when mark = TRUE.}

\item{title}{Logical. If TRUE, title each weblogo.}

\item{title.color}{The color of title of each weblogo}

\item{text_angle}{Numeric. Site number angle.Default:0.}

\item{site_size}{Numeric. Site size.Default:10.}
}
\value{
weblogo for specific site of each family.
}
\description{
weblogo for specific site of each family
}
