% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{reconstruct_2}
\alias{reconstruct_2}
\title{Getting circuit with min dist.}
\usage{
reconstruct_2(ED)
}
\arguments{
\item{ED}{List contain graphical structure.}
}
\value{
circuit with min dist.
}
\description{
Getting circuit with min dist.
}
