% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{relabelgraph}
\alias{relabelgraph}
\title{obtaining relabelled graphs.}
\usage{
relabelgraph(ED)
}
\arguments{
\item{ED}{List contain graphical structure.}
}
\value{
returns the new graphical structure and the mapping between old and new graphs.
}
\description{
obtaining relabelled graphs.
}
