% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Seurat_monocle_workflow.R
\name{Seurat_workflow}
\alias{Seurat_workflow}
\title{Create a Seurat object from score matrix, then, dimensionality reduction and
classification of the score matrix.}
\usage{
Seurat_workflow(
  score_matrix,
  nfeatures = 2000,
  dims_SNN = 1:5,
  resolution = 0.1,
  dims_UMAP = 1:5,
  pca.reduction.name = "pca",
  pca.reduction.key = "PC_",
  umap.reduction.key = "UMAP_",
  ...
)
}
\arguments{
\item{score_matrix}{Matrix. Score matrix of each site.}

\item{nfeatures}{Number of features to select as top variable features.}

\item{dims_SNN}{Dimensions of reduction to use as input with SNN.}

\item{resolution}{Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities.}

\item{dims_UMAP}{Dimensions of reduction to use as input with UMAP.}

\item{pca.reduction.name}{dimensional reduction name with PCA, Default:"pca".}

\item{pca.reduction.key}{dimensional reduction key, specifies the string before the number for the dimension names with PCA. Default:"PC_".}

\item{umap.reduction.key}{dimensional reduction key, specifies the string before the number for the dimension names with UMAP. Default:"UMAP_".}

\item{...}{Other parameters of CreateSeuratObject, FindVariableFeatures, ScaleData, RunPCA, FindNeighbors, FindClusters and RunUMAP functions in Seurat package.}
}
\value{
score matrix of each site. Rownames of the repeated sequence name will be suffixed.
}
\description{
Create a Seurat object from score matrix, then, dimensionality reduction and
classification of the score matrix.
}
\examples{
require(ggseqlogo)
data(ggseqlogo_sample)
test_aa <- get_alignment_score(alignment = seqs_aa[[1]],type = "AA")
my_Seurat <- Seurat_workflow(test_aa)
}
\seealso{
\code{\link{CreateSeuratObject},\link{FindVariableFeatures},\link{ScaleData},
               \link{RunPCA},\link{FindNeighbors},\link{FindClusters},\link{RunUMAP}}
}
