% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{theme_scatter}
\alias{theme_scatter}
\title{Theme of Scatter plot
genepro custom theme}
\usage{
theme_scatter(base_size = 12, base_family = "")
}
\arguments{
\item{base_size}{font base size}

\item{base_family}{font base family}
}
\description{
Theme of Scatter plot
genepro custom theme
}
